Database.Tagger.Type

Description

Connection Types

newtype BareConnection

A newtype that wraps a Connection to create instances of Show and Eq.

Constructors

BareConnection

Fields

_bareConnection :: Connection

data TaggedConnection

A data type that manages connections to a Tagger database.

Constructors

TaggedConnection

Fields

_taggedconnectionConnName :: !Text

Connection name, typically the absolute path of the file connected to, but could be anything.

_taggedconnectionConnInstance :: !(Maybe BareConnection)

Connection to a database, if it is open.

Database Types

type RecordKey = Int

data File

Data type representing a single row in the File table.

Constructors

File

Fields

fileId :: RecordKey

Unique primary key identifying a File record.

filePath :: Text

Unique absolute path to a File.

data Descriptor

Data type corresponding to a single row in the Descriptor table.

Constructors

Descriptor

Fields

descriptorId :: RecordKey

Unique Primary key identifying a Descriptor record.

descriptor :: Text

Descriptor text.

data Tag

Data type corresponding to a single row in the Tag table.

Constructors

Tag

Fields

tagId :: RecordKey

Unique id identifying a single tag record.

tagFileId :: RecordKey

Foreign key to a File record.

tagDescriptorId :: RecordKey

Foreign key to a Descriptor record.

tagSubtagOfId :: Maybe RecordKey

Linking key to another Tag record if this tag is a subtag.

isSubTag :: Tag -> Bool

isSubTagOf :: Tag -> Tag -> Bool

Typically used infix.

data MetaDescriptor

Data type corresponding to one row in the MetaDescriptor Table.

Constructors

MetaDescriptor

Fields

metaDescriptorId :: RecordKey

Foreign key to a Descriptor record.

infraDescriptorId :: RecordKey

Foreign key to a Descriptor record.

Lenses

module Database.Tagger.Type.Lens